requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
Django==5.1.7
djangorestframework==3.16.0
idna==3.10
iniconfig==2.1.0
jwcrypto==1.5.5
packaging==24.2
pluggy==1.5.0
pycparser==2.22
PyJWT==2.10.1
pytest==8.1.1
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
requests==2.32.0
responses==0.23.3
sqlparse==0.5.3
types-PyYAML==6.0.12.20250326
typing_extensions==4.13.0
urllib3==2.3.0
